# tss-website/src/app/api/v1/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TSS API",
    "description": "Universal API for Two Steps Studio - enables external AI agents, applications, and integrations to access TSS data securely.",
    "version": "1.0.0",
    "contact": {
      "name": "Two Steps Studio",
      "url": "https://twostepsstudio.com"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://api.twostepsstudio.com/api/v1",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "API key authentication using Bearer token format: Authorization: Bearer tss_live_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
      }
    },
    "schemas": {
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Project ID"
          },
          "owner_id": {
            "type": "string",
            "description": "Owner user ID"
          },
          "name": {
            "type": "string",
            "description": "Project name"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Project description"
          },
          "description_markdown": {
            "type": "string",
            "nullable": true,
            "description": "Project description in markdown format"
          },
          "color": {
            "type": "string",
            "description": "Project color hex code"
          },
          "status": {
            "type": "string",
            "enum": ["active", "archived"],
            "description": "Project status"
          },
          "project_type": {
            "type": "string",
            "enum": ["general", "web_application", "mobile_application", "desktop_application", "game", "api", "library", "plugin", "other"],
            "description": "Project type"
          },
          "planned_end_date": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Planned end date"
          },
          "is_archived": {
            "type": "boolean",
            "description": "Whether project is archived"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Last update timestamp"
          },
          "user_role": {
            "type": "string",
            "enum": ["owner", "admin", "developer", "tester", "viewer"],
            "description": "User's role in this project"
          },
          "user_permissions": {
            "type": "object",
            "description": "User's permissions in this project",
            "properties": {
              "view_project": { "type": "boolean" },
              "edit_project": { "type": "boolean" },
              "manage_tasks": { "type": "boolean" },
              "manage_roadmap": { "type": "boolean" },
              "manage_files": { "type": "boolean" },
              "manage_description": { "type": "boolean" },
              "manage_technologies": { "type": "boolean" },
              "manage_members": { "type": "boolean" },
              "delete_project": { "type": "boolean" }
            }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Error code"
              },
              "message": {
                "type": "string",
                "description": "Error message"
              },
              "details": {
                "type": "object",
                "description": "Additional error details"
              }
            }
          }
        }
      },
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {}
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer",
                "description": "Current page number"
              },
              "limit": {
                "type": "integer",
                "description": "Items per page"
              },
              "total": {
                "type": "integer",
                "description": "Total number of items"
              },
              "hasMore": {
                "type": "boolean",
                "description": "Whether there are more pages"
              }
            }
          }
        }
      }
    }
  },
  "paths": {
    "/projects": {
      "get": {
        "summary": "List projects",
        "description": "Get all projects accessible to the API key owner. Returns projects where the user is owner or member.",
        "operationId": "listProjects",
        "tags": ["Projects"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page (default: 50, max: 100)",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "includeArchived",
            "in": "query",
            "description": "Include archived projects (default: false)",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests - Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds until retry is allowed",
                "schema": {
                  "type": "integer"
                }
              }
            }
          }
        },
        "x-required-scopes": ["projects:read"]
      },
      "post": {
        "summary": "Create project",
        "description": "Create a new project. The user must be within their project limit.",
        "operationId": "createProject",
        "tags": ["Projects"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Project name"
                  },
                  "description": {
                    "type": "string",
                    "description": "Project description"
                  },
                  "description_markdown": {
                    "type": "string",
                    "description": "Project description in markdown format"
                  },
                  "color": {
                    "type": "string",
                    "description": "Project color hex code (default: #ffcb2f)"
                  },
                  "project_type": {
                    "type": "string",
                    "enum": ["general", "web_application", "mobile_application", "desktop_application", "game", "api", "library", "plugin", "other"],
                    "description": "Project type (default: general)"
                  },
                  "planned_end_date": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Planned end date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input or project limit reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests - Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds until retry is allowed",
                "schema": {
                  "type": "integer"
                }
              }
            }
          }
        },
        "x-required-scopes": ["projects:write"]
      }
    },
...
  },
...
}